+++ test/test_mqtt.py
@@ -1,16 +1,17 @@
 # -*- coding: utf-8 -*-
 # (c) 2020 Andreas Motl <andreas@getkotori.org>
 import logging
+
 import pytest_twisted
 
-from test.resources import influx, mqtt_topic
+from test.resources import influx, mqtt_topic, PROCESS_DELAY
 from test.util import mqtt_sensor, sleep
 
 logger = logging.getLogger(__name__)
 
 
 @pytest_twisted.inlineCallbacks
-def test_mqtt_to_influxdb_basic(machinery, create_influxdb, reset_influxdb):
+def test_mqtt_to_influxdb(machinery, create_influxdb, reset_influxdb):
 
     # Submit a single measurement, without timestamp.
     data = {
@@ -20,9 +21,9 @@ def test_mqtt_to_influxdb_basic(machinery, create_influxdb, reset_influxdb):
     yield mqtt_sensor(mqtt_topic, data)
 
     # Wait for some time to process the message.
-    yield sleep(0.2)
+    yield sleep(PROCESS_DELAY)
 
-    # Check if data arrived in InfluxDB.
+    # Proof that data arrived in InfluxDB.
     record = influx.get_first_record()
     del record['time']
     assert record == {u'humidity': 83.1, u'temperature': 42.84}