+++ CHANGES.rst
@@ -8,7 +8,8 @@ in progress
 - Improve Docker image building
 - Add packaging for Ubuntu 18 Bionic Beaver
 - Update documentation, now using "Material for Sphinx" theme
-- Add test harness
+- Add InfluxDB integration tests
+- Add Grafana integration tests
 
 
 .. _kotori-0.24.5: